support transfer object.
