+++ src/worker/model/asset.ts
@@ -4,6 +4,12 @@ export interface UploadAssetMessage {
     encrypted: boolean,
 }
 
+export interface DownloadAssetMessage {
+    namespaceId: string,
+    hash: string,
+    merged: boolean,
+}
+
 export function getClientImports() {
     return [`import { Uploader, Downloader } from '${getModulePath('@yeying-community/yeying-client-ts')}';`]
 }
